pyproject.toml
[tool.poetry]
name = "gtaradiobot"
version = "0.3.0"
description = "Discord GTA radio bot"
authors = ["WWatts"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
aiohttp = "3.7.4.post0"
async-timeout = "3.0.1"
atomicwrites = "1.4.0"
attrs = "21.2.0"
black = "^22.10.0"
cffi = "1.15.0"
chardet = "4.0.0"
click = "8.0.3"
colorama = "0.4.4"
"discord.py" = "^2.0.1"
fuzzywuzzy = "0.18.0"
idna = "3.3"
iniconfig = "1.1.1"
multidict = "5.2.0"
mutagen = "1.45.1"
mypy-extensions = "0.4.3"
packaging = "21.3"
pathspec = "0.9.0"
platformdirs = "2.4.1"
pluggy = "1.0.0"
py = "1.11.0"
pycparser = "2.21"
PyNaCl = "^1.5.0"
pyparsing = "3.0.6"
pytest = "6.2.5"
python-Levenshtein = "0.12.2"
six = "1.16.0"
toml = "0.10.2"
tomli = "1.2.3"
typing-extensions = "4.0.1"
yarl = "1.7.2"
nest-asyncio = "^1.5.6"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
